+++ src/onegov.core/onegov/core/tests/test_locking.py
@@ -9,13 +9,16 @@ from threading import Thread
 
 class LockingThread(Thread):
 
-    def __init__(self, mgr, namespace, key):
+    def __init__(self, mgr, namespace, key, schema):
         Thread.__init__(self)
         self.mgr = mgr
         self.namespace = namespace
         self.key = key
+        self.schema = schema
 
     def run(self):
+        self.mgr.set_current_schema(self.schema)
+
         try:
             with lock(self.mgr.session(), self.namespace, self.key):
                 time.sleep(0.1)
@@ -37,8 +40,30 @@ def test_threaded_locking(postgres_dsn):
     session_manager.set_current_schema('foo')
 
     threads = [
-        LockingThread(session_manager, 'namespace', 'key'),
-        LockingThread(session_manager, 'namespace', 'key')
+        LockingThread(session_manager, 'namespace', 'key', 'foo'),
+        LockingThread(session_manager, 'namespace', 'key', 'foo')
+    ]
+
+    for t in threads:
+        t.start()
+
+    for t in threads:
+        t.join()
+
+    exceptions = [t.exception for t in threads if t.exception]
+
+    assert len(exceptions) == 1
+    assert isinstance(exceptions[0], AlreadyLockedError)
+
+
+def test_multischema_threaded_locking(postgres_dsn):
+    Base = declarative_base()
+
+    session_manager = SessionManager(postgres_dsn, Base)
+
+    threads = [
+        LockingThread(session_manager, 'namespace', 'key', 'foo'),
+        LockingThread(session_manager, 'namespace', 'key', 'bar')
     ]
 
     for t in threads: